Improve Streamlit UI
